factor out graphql types into separate classes, update to schema
